requirements.txt
# Build dependencies
hatchling>=1.21.0

# Core dependencies
dramatiq[rabbitmq]>=1.17.1
Flask>=3.0.0
flask-cors==3.0.10
python-dotenv>=1.0.0

# Database dependencies
pymongo>=4.6.0
pika>=1.3.0
redis>=5.0.0

# Video processing dependencies
yt-dlp>=2023.12.30
moviepy>=1.0.3
pytube>=15.0.0
openai>=1.3.0
boto3>=1.34.0
watchdog>=3.0.0

# Web scraping dependencies
undetected-chromedriver>=3.5.3
beautifulsoup4>=4.12.2
selenium>=4.15.2

# Utility dependencies
requests>=2.31.0
python-dateutil>=2.8.2
tqdm>=4.66.0

# Organization packages
-e utom_databases
-e utom_utils

# Additional dependencies
ffmpeg-python>=0.2.0
spacy>=3.7.2 